+++ src/JsonApi/Client/JsonApiClient.php
@@ -1,9 +1,9 @@
 <?php
-namespace WoohooLabs\Yang\JsonApi;
+namespace WoohooLabs\Yang\JsonApi\Client;
 
 use Http\Client\HttpClient;
 use Psr\Http\Message\RequestInterface;
-use WoohooLabs\Yang\JsonApi\Deserializer\DefaultDeserializer;
+use WoohooLabs\Yang\JsonApi\Deserializer\JsonDeserializer;
 use WoohooLabs\Yang\JsonApi\Deserializer\DeserializerInterface;
 use WoohooLabs\Yang\JsonApi\Response\JsonApiResponse;
 
@@ -22,7 +22,7 @@ class JsonApiClient
     public function __construct(HttpClient $client, DeserializerInterface $deserializer = null)
     {
         $this->client = $client;
-        $this->deserializer = $deserializer ? $deserializer : new DefaultDeserializer();
+        $this->deserializer = $deserializer ? $deserializer : new JsonDeserializer();
     }
 
     /**